% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openHealthCanada.R
\name{recent_recalls}
\alias{recent_recalls}
\title{Recent Recalls}
\usage{
recent_recalls(lang = "en", tibble = TRUE)
}
\description{
Returns a response with the 15 latest recalls in each category,
all, food, vehicles, health, and cps.

 @param lang Either \code{en} for english or \code{fr} for french, defaults
 to \code{en}.
 @param tibble Logical \code{TRUE} or \code{FALSE} to return a tibble,
 defaults to \code{TRUE}.

 @example
 x <- recent_recalls()
 str(x, max.level = 1)

 @export
}
